requirements.txt
mcp[cli]>=1.6.0
psycopg2-binary>=2.9.5
python-dotenv>=1.0.0
langchain>=0.3.24
langchain-groq>=0.3.2